--- src/main/java/com/khomishchak/ws/repositories/custom/ApiKeySettingCustomRepository.java
@@ -1,9 +1,0 @@
-package com.khomishchak.ws.repositories.custom;
-
-import com.khomishchak.ws.model.exchanger.DecryptedApiKeySettingDTO;
-
-import java.util.List;
-
-public interface ApiKeySettingCustomRepository {
-    List<DecryptedApiKeySettingDTO> findAllByUserIdDecrypted(long userId);
-}
