+++ src/main/java/com/khomishchak/ws/services/exchangers/apikeys/ApiKeySettingService.java
@@ -1,0 +1,9 @@
+package com.khomishchak.ws.services.exchangers.apikeys;
+
+import com.khomishchak.ws.model.exchanger.DecryptedApiKeySettingDTO;
+
+import java.util.List;
+
+public interface ApiKeySettingService {
+    List<DecryptedApiKeySettingDTO> getDecryptApiKeySettings(long userId);
+}
